+++ packages/devextreme/js/__internal/core/license/key.ts
@@ -1,0 +1,17 @@
+export interface PublicKey {
+  e: number;
+  n: Uint8Array;
+}
+
+export const PUBLIC_KEY: PublicKey = {
+  e: 65537,
+  n: new Uint8Array([
+    202, 208, 20, 244, 235, 89, 121, 253, 219, 161, 162, 26, 166, 22, 65, 81, 176, 0, 101, 246,
+    34, 101, 128, 51, 224, 52, 194, 227, 113, 10, 4, 96, 201, 33, 171, 251, 204, 57, 164, 28, 89,
+    249, 191, 46, 170, 74, 37, 125, 216, 95, 240, 125, 69, 31, 134, 79, 101, 62, 25, 30, 162, 31,
+    206, 104, 92, 42, 35, 164, 93, 97, 197, 198, 239, 225, 249, 146, 119, 88, 20, 76, 219, 218,
+    113, 0, 29, 246, 132, 116, 37, 252, 113, 87, 200, 99, 171, 146, 136, 182, 216, 226, 97, 67,
+    85, 126, 103, 117, 236, 49, 60, 32, 109, 91, 139, 166, 1, 152, 228, 36, 182, 167, 19, 106,
+    72, 62, 186, 243, 199, 73,
+  ]),
+};
